+++ tests-integration/tests/kernel_orch.rs
@@ -1,4 +1,5 @@
 use andromeda_adodb::ADODBContract;
+use andromeda_auction::{mock::mock_start_auction, AuctionContract};
 use andromeda_counter::CounterContract;
 use andromeda_economics::EconomicsContract;
 use andromeda_finance::splitter::{
@@ -9,24 +10,24 @@ use andromeda_math::counter::{
     InstantiateMsg as CounterInstantiateMsg, State,
 };
 
+use andromeda_cw721::CW721Contract;
 use andromeda_kernel::KernelContract;
+use andromeda_non_fungible_tokens::cw721::TokenExtension;
 use andromeda_splitter::SplitterContract;
 use andromeda_std::{
+    ado_base::rates::{LocalRate, LocalRateType, LocalRateValue, PercentRate, Rate, RatesMessage},
     amp::{
         messages::{AMPMsg, AMPMsgConfig},
         AndrAddr, Recipient,
     },
-    common::Milliseconds,
+    common::{denom::Asset, expiration::Expiry, Milliseconds},
     os::{
         self,
-        kernel::{AcknowledgementMsg, ExecuteMsg, InstantiateMsg, SendMessageWithFundsResponse},
+        kernel::{ExecuteMsg, InstantiateMsg},
     },
 };
 use andromeda_vfs::VFSContract;
-use cosmwasm_std::{
-    to_json_binary, Addr, Binary, Decimal, IbcAcknowledgement, IbcEndpoint, IbcPacket,
-    IbcPacketAckMsg, IbcTimeout, Timestamp, Uint128,
-};
+use cosmwasm_std::{coin, to_json_binary, Addr, Binary, Decimal, StdAck, Uint128};
 use cw_orch::prelude::*;
 use cw_orch_interchain::{prelude::*, types::IbcPacketOutcome, InterchainEnv};
 use ibc_relayer_types::core::ics24_host::identifier::PortId;
@@ -586,6 +587,7 @@ fn test_kernel_ibc_execute_only_multi_hop() {
 fn test_kernel_ibc_funds_only() {
     // Here `juno-1` is the chain-id and `juno` is the address prefix for this chain
     let sender = Addr::unchecked("sender_for_all_chains").into_string();
+    let buyer = Addr::unchecked("buyer").into_string();
 
     let interchain = MockInterchainEnv::new(vec![
         ("juno", &sender),
@@ -596,12 +598,17 @@ fn test_kernel_ibc_funds_only() {
 
     let juno = interchain.get_chain("juno").unwrap();
     let osmosis = interchain.get_chain("osmosis").unwrap();
-
     juno.set_balance(sender.clone(), vec![Coin::new(100000000000000, "juno")])
+        .unwrap();
+    juno.set_balance(buyer.clone(), vec![Coin::new(100000000000000, "juno")])
         .unwrap();
 
     let kernel_juno = KernelContract::new(juno.clone());
     let vfs_juno = VFSContract::new(juno.clone());
+    let adodb_juno = ADODBContract::new(juno.clone());
+    let economics_juno = EconomicsContract::new(juno.clone());
+    let mut auction_juno = AuctionContract::new(juno.clone());
+    let cw721_juno = CW721Contract::new(juno.clone());
     let kernel_osmosis = KernelContract::new(osmosis.clone());
     let counter_osmosis = CounterContract::new(osmosis.clone());
     let vfs_osmosis = VFSContract::new(osmosis.clone());
@@ -610,6 +617,11 @@ fn test_kernel_ibc_funds_only() {
 
     kernel_juno.upload().unwrap();
     vfs_juno.upload().unwrap();
+    adodb_juno.upload().unwrap();
+    economics_juno.upload().unwrap();
+    auction_juno.upload().unwrap();
+    cw721_juno.upload().unwrap();
+
     kernel_osmosis.upload().unwrap();
     counter_osmosis.upload().unwrap();
     vfs_osmosis.upload().unwrap();
@@ -703,6 +715,38 @@ fn test_kernel_ibc_funds_only() {
         )
         .unwrap();
 
+    adodb_juno
+        .instantiate(
+            &os::adodb::InstantiateMsg {
+                kernel_address: kernel_juno.address().unwrap().into_string(),
+                owner: None,
+            },
+            None,
+            None,
+        )
+        .unwrap();
+
+    economics_juno
+        .instantiate(
+            &os::economics::InstantiateMsg {
+                kernel_address: kernel_juno.address().unwrap().into_string(),
+                owner: None,
+            },
+            None,
+            None,
+        )
+        .unwrap();
+
+    kernel_juno
+        .execute(
+            &ExecuteMsg::UpsertKeyAddress {
+                key: "economics".to_string(),
+                value: economics_juno.address().unwrap().into_string(),
+            },
+            None,
+        )
+        .unwrap();
+
     adodb_osmosis
         .instantiate(
             &os::adodb::InstantiateMsg {
@@ -714,14 +758,11 @@ fn test_kernel_ibc_funds_only() {
         )
         .unwrap();
 
-    adodb_osmosis
+    kernel_juno
         .execute(
-            &os::adodb::ExecuteMsg::Publish {
-                code_id: 2,
-                ado_type: "counter".to_string(),
-                action_fees: None,
-                version: "1.0.2".to_string(),
-                publisher: None,
+            &ExecuteMsg::UpsertKeyAddress {
+                key: "vfs".to_string(),
+                value: vfs_juno.address().unwrap().into_string(),
             },
             None,
         )
@@ -730,8 +771,8 @@ fn test_kernel_ibc_funds_only() {
     kernel_juno
         .execute(
             &ExecuteMsg::UpsertKeyAddress {
-                key: "vfs".to_string(),
-                value: vfs_juno.address().unwrap().into_string(),
+                key: "adodb".to_string(),
+                value: adodb_juno.address().unwrap().into_string(),
             },
             None,
         )
@@ -772,7 +813,7 @@ fn test_kernel_ibc_funds_only() {
     kernel_osmosis
         .execute(
             &ExecuteMsg::AssignChannels {
-                ics20_channel_id: Some(channel.0.channel.unwrap().to_string()),
+                ics20_channel_id: Some(channel.0.channel.clone().unwrap().to_string()),
                 direct_channel_id: Some(juno_channel.to_string()),
                 chain: "juno".to_string(),
                 kernel_address: kernel_juno.address().unwrap().into_string(),
@@ -782,6 +823,10 @@ fn test_kernel_ibc_funds_only() {
         .unwrap();
 
     let recipient = "osmo1qzskhrca90qy2yjjxqzq4yajy842x7c50xq33d";
+    println!(
+        "osmosis kernel address: {}",
+        kernel_osmosis.address().unwrap()
+    );
 
     let kernel_juno_send_request = kernel_juno
         .execute(
@@ -836,7 +881,7 @@ fn test_kernel_ibc_funds_only() {
         .execute(
             &ExecuteMsg::UpsertKeyAddress {
                 key: "trigger_key".to_string(),
-                value: sender,
+                value: sender.clone(),
             },
             None,
         )
@@ -846,29 +891,9 @@ fn test_kernel_ibc_funds_only() {
     let kernel_juno_trigger_request = kernel_juno
         .execute(
             &ExecuteMsg::TriggerRelay {
-                packet_sequence: "1".to_string(),
-                packet_ack_msg: IbcPacketAckMsg::new(
-                    IbcAcknowledgement::new(
-                        to_json_binary(&AcknowledgementMsg::<SendMessageWithFundsResponse>::Ok(
-                            SendMessageWithFundsResponse {},
-                        ))
-                        .unwrap(),
-                    ),
-                    IbcPacket::new(
-                        Binary::default(),
-                        IbcEndpoint {
-                            port_id: "port_id".to_string(),
-                            channel_id: "channel_id".to_string(),
-                        },
-                        IbcEndpoint {
-                            port_id: "port_id".to_string(),
-                            channel_id: "channel_id".to_string(),
-                        },
-                        1,
-                        IbcTimeout::with_timestamp(Timestamp::from_seconds(1)),
-                    ),
-                    Addr::unchecked("relayer"),
-                ),
+                packet_sequence: 1,
+                channel_id: channel.0.channel.clone().unwrap().to_string(),
+                packet_ack: to_json_binary(&StdAck::Success(Binary::default())).unwrap(),
             },
             None,
         )
@@ -887,6 +912,162 @@ fn test_kernel_ibc_funds_only() {
         assert_eq!(balances.len(), 1);
         assert_eq!(balances[0].denom, ibc_denom);
         assert_eq!(balances[0].amount.u128(), 100);
+    } else {
+        panic!("packet timed out");
+        // There was a decode error or the packet timed out
+        // Else the packet timed-out, you may have a relayer error or something is wrong in your application
+    };
+
+    // Set up cross chain rates recipient
+    auction_juno
+        .instantiate(
+            &andromeda_non_fungible_tokens::auction::InstantiateMsg {
+                authorized_token_addresses: None,
+                authorized_cw20_addresses: None,
+                kernel_address: kernel_juno.address().unwrap().into_string(),
+                owner: None,
+            },
+            None,
+            None,
+        )
+        .unwrap();
+
+    cw721_juno
+        .instantiate(
+            &andromeda_non_fungible_tokens::cw721::InstantiateMsg {
+                name: "test tokens".to_string(),
+                symbol: "TT".to_string(),
+                minter: AndrAddr::from_string(sender.clone()),
+                kernel_address: kernel_juno.address().unwrap().into_string(),
+                owner: None,
+            },
+            None,
+            None,
+        )
+        .unwrap();
+
+    auction_juno
+        .execute(
+            &andromeda_non_fungible_tokens::auction::ExecuteMsg::Rates(RatesMessage::SetRate {
+                action: "Claim".to_string(),
+                rate: Rate::Local(LocalRate {
+                    rate_type: LocalRateType::Deductive,
+                    recipient: Recipient::new(
+                        AndrAddr::from_string(format!("ibc://osmosis/{}", recipient)),
+                        None,
+                    ),
+                    value: LocalRateValue::Percent(PercentRate {
+                        percent: Decimal::percent(50),
+                    }),
+                    description: None,
+                }),
+            }),
+            None,
+        )
+        .unwrap();
+
+    cw721_juno
+        .execute(
+            &andromeda_non_fungible_tokens::cw721::ExecuteMsg::Mint {
+                token_id: "1".to_string(),
+                owner: sender.clone(),
+                token_uri: None,
+                extension: TokenExtension::default(),
+            },
+            None,
+        )
+        .unwrap();
+
+    let start_time = Milliseconds::from_nanos(juno.block_info().unwrap().time.nanos());
+    let receive_msg = mock_start_auction(
+        None,
+        Expiry::AtTime(start_time.plus_milliseconds(Milliseconds(10000))),
+        None,
+        Asset::NativeToken("juno".to_string()),
+        None,
+        None,
+        None,
+        None,
+    );
+    cw721_juno
+        .execute(
+            &andromeda_non_fungible_tokens::cw721::ExecuteMsg::SendNft {
+                contract: AndrAddr::from_string(auction_juno.address().unwrap()),
+                token_id: "1".to_string(),
+                msg: to_json_binary(&receive_msg).unwrap(),
+            },
+            None,
+        )
+        .unwrap();
+    juno.wait_seconds(1).unwrap();
+
+    auction_juno.set_sender(&Addr::unchecked(buyer.clone()));
+    auction_juno
+        .execute(
+            &andromeda_non_fungible_tokens::auction::ExecuteMsg::PlaceBid {
+                token_id: "1".to_string(),
+                token_address: cw721_juno.address().unwrap().into_string(),
+            },
+            Some(&[coin(50, "juno")]),
+        )
+        .unwrap();
+    juno.next_block().unwrap();
+    juno.next_block().unwrap();
+
+    // Claim
+    let claim_request = auction_juno
+        .execute(
+            &andromeda_non_fungible_tokens::auction::ExecuteMsg::Claim {
+                token_id: "1".to_string(),
+                token_address: cw721_juno.address().unwrap().into_string(),
+            },
+            None,
+        )
+        .unwrap();
+    let packet_lifetime = interchain.await_packets("juno", claim_request).unwrap();
+
+    // For testing a successful outcome of the first packet sent out in the tx, you can use:
+    if let IbcPacketOutcome::Success { .. } = &packet_lifetime.packets[0].outcome {
+        // Packet has been successfully acknowledged and decoded, the transaction has gone through correctly
+
+        // Check recipient balance after trigger execute msg
+        let balances = osmosis
+            .query_all_balances(kernel_osmosis.address().unwrap())
+            .unwrap();
+        assert_eq!(balances.len(), 1);
+        assert_eq!(balances[0].denom, ibc_denom);
+        assert_eq!(balances[0].amount.u128(), 25);
+    } else {
+        panic!("packet timed out");
+        // There was a decode error or the packet timed out
+        // Else the packet timed-out, you may have a relayer error or something is wrong in your application
+    };
+
+    // Construct an Execute msg from the kernel on juno inteded for the splitter on osmosis
+    let kernel_juno_trigger_request = kernel_juno
+        .execute(
+            &ExecuteMsg::TriggerRelay {
+                packet_sequence: 2,
+                channel_id: channel.0.channel.clone().unwrap().to_string(),
+                packet_ack: to_json_binary(&StdAck::Success(Binary::default())).unwrap(),
+            },
+            None,
+        )
+        .unwrap();
+
+    let packet_lifetime = interchain
+        .await_packets("juno", kernel_juno_trigger_request)
+        .unwrap();
+
+    // For testing a successful outcome of the first packet sent out in the tx, you can use:
+    if let IbcPacketOutcome::Success { .. } = &packet_lifetime.packets[0].outcome {
+        // Packet has been successfully acknowledged and decoded, the transaction has gone through correctly
+
+        // Check recipient balance after trigger execute msg
+        let balances = osmosis.query_all_balances(recipient).unwrap();
+        assert_eq!(balances.len(), 1);
+        assert_eq!(balances[0].denom, ibc_denom);
+        assert_eq!(balances[0].amount.u128(), 100 + 25);
     } else {
         panic!("packet timed out");
         // There was a decode error or the packet timed out
@@ -1008,19 +1189,6 @@ fn test_kernel_ibc_funds_only_multi_hop() {
                 owner: None,
             },
             None,
-            None,
-        )
-        .unwrap();
-
-    adodb_osmosis
-        .execute(
-            &os::adodb::ExecuteMsg::Publish {
-                code_id: 2,
-                ado_type: "counter".to_string(),
-                action_fees: None,
-                version: "1.0.2".to_string(),
-                publisher: None,
-            },
             None,
         )
         .unwrap();
@@ -1405,19 +1573,6 @@ fn test_kernel_ibc_funds_and_execute_msg() {
         )
         .unwrap();
 
-    adodb_osmosis
-        .execute(
-            &os::adodb::ExecuteMsg::Publish {
-                code_id: 2,
-                ado_type: "counter".to_string(),
-                action_fees: None,
-                version: "1.0.2".to_string(),
-                publisher: None,
-            },
-            None,
-        )
-        .unwrap();
-
     kernel_juno
         .execute(
             &ExecuteMsg::UpsertKeyAddress {
@@ -1473,7 +1628,7 @@ fn test_kernel_ibc_funds_and_execute_msg() {
     kernel_osmosis
         .execute(
             &ExecuteMsg::AssignChannels {
-                ics20_channel_id: Some(channel.0.channel.unwrap().to_string()),
+                ics20_channel_id: Some(channel.0.channel.clone().unwrap().to_string()),
                 direct_channel_id: Some(juno_channel.to_string()),
                 chain: "juno".to_string(),
                 kernel_address: kernel_juno.address().unwrap().into_string(),
@@ -1565,31 +1720,9 @@ fn test_kernel_ibc_funds_and_execute_msg() {
         let kernel_juno_splitter_request = kernel_juno
             .execute(
                 &ExecuteMsg::TriggerRelay {
-                    packet_sequence: "1".to_string(),
-                    packet_ack_msg: IbcPacketAckMsg::new(
-                        IbcAcknowledgement::new(
-                            to_json_binary(
-                                &AcknowledgementMsg::<SendMessageWithFundsResponse>::Ok(
-                                    SendMessageWithFundsResponse {},
-                                ),
-                            )
-                            .unwrap(),
-                        ),
-                        IbcPacket::new(
-                            Binary::default(),
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            1,
-                            IbcTimeout::with_timestamp(Timestamp::from_seconds(1)),
-                        ),
-                        Addr::unchecked("relayer"),
-                    ),
+                    packet_sequence: 1,
+                    channel_id: channel.0.channel.clone().unwrap().to_string(),
+                    packet_ack: to_json_binary(&StdAck::Success(Binary::default())).unwrap(),
                 },
                 None,
             )
@@ -1795,7 +1928,7 @@ fn test_kernel_ibc_funds_only_unhappy() {
     kernel_osmosis
         .execute(
             &ExecuteMsg::AssignChannels {
-                ics20_channel_id: Some(channel.0.channel.unwrap().to_string()),
+                ics20_channel_id: Some(channel.0.channel.clone().unwrap().to_string()),
                 direct_channel_id: Some(juno_channel.to_string()),
                 chain: "juno".to_string(),
                 kernel_address: kernel_juno.address().unwrap().into_string(),
@@ -1860,31 +1993,9 @@ fn test_kernel_ibc_funds_only_unhappy() {
         let kernel_juno_splitter_request = kernel_juno
             .execute(
                 &ExecuteMsg::TriggerRelay {
-                    packet_sequence: "1".to_string(),
-                    packet_ack_msg: IbcPacketAckMsg::new(
-                        IbcAcknowledgement::new(
-                            to_json_binary(
-                                &AcknowledgementMsg::<SendMessageWithFundsResponse>::Error(
-                                    "error".to_string(),
-                                ),
-                            )
-                            .unwrap(),
-                        ),
-                        IbcPacket::new(
-                            Binary::default(),
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            1,
-                            IbcTimeout::with_timestamp(Timestamp::from_seconds(1)),
-                        ),
-                        Addr::unchecked("relayer"),
-                    ),
+                    packet_sequence: 1,
+                    channel_id: channel.0.channel.clone().unwrap().to_string(),
+                    packet_ack: to_json_binary(&StdAck::Error("error".to_string())).unwrap(),
                 },
                 None,
             )
@@ -2117,7 +2228,7 @@ fn test_kernel_ibc_funds_and_execute_msg_unhappy() {
     kernel_osmosis
         .execute(
             &ExecuteMsg::AssignChannels {
-                ics20_channel_id: Some(channel.0.channel.unwrap().to_string()),
+                ics20_channel_id: Some(channel.0.channel.clone().unwrap().to_string()),
                 direct_channel_id: Some(juno_channel.to_string()),
                 chain: "juno".to_string(),
                 kernel_address: kernel_juno.address().unwrap().into_string(),
@@ -2213,32 +2324,9 @@ fn test_kernel_ibc_funds_and_execute_msg_unhappy() {
         let kernel_juno_splitter_request = kernel_juno
             .execute(
                 &ExecuteMsg::TriggerRelay {
-                    packet_sequence: "1".to_string(),
-                    packet_ack_msg: IbcPacketAckMsg::new(
-                        IbcAcknowledgement::new(
-                            to_json_binary(
-                                // It's Ok because the ics20 transfer is supposed to go through. We want the ExecuteMsg to fail
-                                &AcknowledgementMsg::<SendMessageWithFundsResponse>::Ok(
-                                    SendMessageWithFundsResponse {},
-                                ),
-                            )
-                            .unwrap(),
-                        ),
-                        IbcPacket::new(
-                            Binary::default(),
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            IbcEndpoint {
-                                port_id: "port_id".to_string(),
-                                channel_id: "channel_id".to_string(),
-                            },
-                            1,
-                            IbcTimeout::with_timestamp(Timestamp::from_seconds(1)),
-                        ),
-                        Addr::unchecked("relayer"),
-                    ),
+                    packet_sequence: 1,
+                    channel_id: channel.0.channel.clone().unwrap().to_string(),
+                    packet_ack: to_json_binary(&StdAck::Success(Binary::default())).unwrap(),
                 },
                 None,
             )